Admin-reports code refactor
